ref: Move CacheKey into its own file
